requirements.txt
fastapi[all]
sqlalchemy
uvicorn
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]